+++ CHANGELOG.md
@@ -16,6 +16,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- イベントのカタログ情報をoffersからhasOfferCatalogへ分離
+
 ### Deprecated
 
 ### Removed